Cleaned build system, refactored directory structure, builds on Mac OS X
