--- Makefile.in
@@ -1,35 +1,0 @@
-.PHONY: clean all install install_udev
-
-export CC = @CC@
-export AR = @AR@
-export CFLAGS = @CPPFLAGS@ @CFLAGS@ -Wall -Wextra -Wno-unused-parameter
-export LDFLAGS = @LDFLAGS@
-export PREFIX = @prefix@
-
-all:
-	$(MAKE) -C ttbin
-	$(MAKE) -C libttwatch
-	$(MAKE) -C ttwatch
-	$(MAKE) -C ttbincnv
-	$(MAKE) -C ttbinmod
-
-clean:
-	$(MAKE) -C ttbin clean
-	$(MAKE) -C libttwatch clean
-	$(MAKE) -C ttwatch clean
-	$(MAKE) -C ttbincnv clean
-	$(MAKE) -C ttbinmod clean
-
-install:
-	$(MAKE) -C ttwatch install
-	$(MAKE) -C ttbincnv install
-	$(MAKE) -C ttbinmod install
-
-install_udev:
-	cp 99-tomtom.rules /etc/udev/rules.d
-	udevadm control --reload-rules
-	-addgroup usb
-	-usermod -a -Gusb `logname`
-
-# vim: noet
-
