+++ include/libttwatch.h
@@ -3,7 +3,7 @@
 **
 ** main header file for the ttwatch library
 *******************************************************************************/
-
+  
 #ifndef __LIBTTWATCH_H__
 #define __LIBTTWATCH_H__
