+++ manifest/make_manifest.pl
