+++ src/ttbincnv.c
